refactored to create title checker class
